+++ src/components/tabbar/tabbar-button.tsx
@@ -1,0 +1,83 @@
+import { useEffect } from 'react';
+import { Pressable, StyleSheet } from 'react-native';
+import Animated, {
+  interpolate,
+  useAnimatedStyle,
+  useSharedValue,
+  withSpring,
+} from 'react-native-reanimated';
+
+import { icon } from '@/constants/icon';
+
+export const TabBarButton = ({
+  onPress,
+  onLongPress,
+  isFocused,
+  routeName,
+  color,
+  label,
+}: {
+  onPress: Function;
+  onLongPress: Function;
+  isFocused: boolean;
+  routeName: string;
+  color: string;
+  label: string;
+}) => {
+  const scale = useSharedValue(0);
+
+  useEffect(() => {
+    scale.value = withSpring(
+      typeof isFocused === 'boolean' ? (isFocused ? 1 : 0) : isFocused,
+      { duration: 350 },
+    );
+  }, [scale, isFocused]);
+
+  const animatedIconStyle = useAnimatedStyle(() => {
+    const scaleValue = interpolate(scale.value, [0, 1], [1, 1.2]);
+    const top = interpolate(scale.value, [0, 1], [0, 9]);
+    return {
+      transform: [
+        {
+          scale: scaleValue,
+        },
+      ],
+      top: top,
+    };
+  });
+
+  const animatedTextStyle = useAnimatedStyle(() => {
+    const opacity = interpolate(scale.value, [0, 1], [1, 0]);
+    return {
+      opacity,
+    };
+  });
+  return (
+    <Pressable
+      onPress={onPress}
+      onLongPress={onLongPress}
+      style={styles.tabbarItem}
+    >
+      <Animated.View style={animatedIconStyle}>
+        {icon[routeName]({ color: color })}
+      </Animated.View>
+
+      <Animated.Text
+        style={[
+          { color: isFocused ? '#673ab7' : '#222', fontSize: 12 },
+          animatedTextStyle,
+        ]}
+      >
+        {label}
+      </Animated.Text>
+    </Pressable>
+  );
+};
+
+const styles = StyleSheet.create({
+  tabbarItem: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
