+++ src/constants/icon.tsx
@@ -1,0 +1,8 @@
+import AntDesign from '@expo/vector-icons/AntDesign';
+
+export const icon = {
+  index: (props: any) => <AntDesign name="home" size={24} {...props} />,
+  play: (props: any) => <AntDesign name="addusergroup" size={24} {...props} />,
+  book: (props: any) => <AntDesign name="creditcard" size={24} {...props} />,
+  settings: (props: any) => <AntDesign name="setting" size={24} {...props} />,
+};
